### 2021-05-03 #55 @mesnardo
[Re] Three-dimensional wake topology and propulsive performance of low-aspect-ratio pitching-rolling plates
Dear Editors,

I would like to request a review for the following replication:

**Original article:** [Three-dimensional wake topology and propulsive performance of low-aspect-ratio pitching-rolling plates](https://doi.org/10.1063/1.4954505)

**PDF URL:** [[Re] Three-dimensional wake topology and propulsive performance of low-aspect-ratio pitching-rolling plates](https://github.com/barbagroup/rescience-rollingpitching/blob/submission-2021.05.02/article.pdf)
**Metadata URL:** [`metadata.yaml`](https://github.com/barbagroup/rescience-rollingpitching/blob/d5ab9c2b82e14354eb92b9226f4cf465646374fc/metadata.yaml)
**Code URL:** [`barbagroup/petibm-rollingpitching`](https://github.com/barbagroup/petibm-rollingpitching/tree/23f93bb8f7f7ccddc37a4d315813308adad7aea9) (doi: [10.5281/zenodo.4733323](https://doi.org/10.5281/zenodo.4733323))

**Scientific domain:** Computational Fluid Dynamics
**Programming language:**  C++ (code application), Python (processing scripts)

cc @labarba (co-author)
