### 2021-05-06 #55 @khinsen
Thanks for this submission @mesnardo, which I think is our first one in computational fluid dynamics. Our only editor in this field is your co-author @labarba, so we will have to improvise a bit.

I can edit this myself, profiting from many contacts in the CFD community, but I won't be able to start before another two weeks due to existing commitments.

@ReScience/editors Can anyone else handle this, ideally more rapidly than myself?
