### 2021-05-26 #55 @labarba
Maybe the author of this published ReScience paper could contribute a peer review here, and thus give back to the efforts of the journal? https://github.com/ReScience/submissions/issues/19
