### 2021-05-27 #55 @labarba
To clarify, Dr. Chengyu Li is the first author of the original article from 2016 (https://doi.org/10.1063/1.4954505) this is a replication of. He was not involved in the replication and we have not interacted before, but I reached out by email this week to ask if he might be interested in reviewing the replication.
