### 2021-05-27 #55 @khinsen
@chengyu1018 Welcome to ReScience, and thanks for volunteering as a reviewer. I can't think of anyone better placed to do a review! If you have any questions about how ReScience functions, don't hesitate to ask. To start, [here](https://rescience.github.io/edit/) are our reviewing guidelines.

Normally, we expect reviewers to verify the reproducibility of the submitted code. This could be difficult in this case because the computations require computing resource that are not on every scientist's desk. See what you can do, and if there's any problem due to the resource requirements, we will adjust. This is our first HPC submission as far as I can see.
