### 2021-05-27 #55 @pantale
@khinsen Hello Konrad,
I would be very happy to take part in this review. I am not an expert in numerical simulation in fluid mechanics, my field of expertise is in solid mechanics, nevertheless, I think that with the help of the first author of the original article I can bring my point of view on this replication even if I have neither the material means nor the expertise to carry out the replication of this computation by myself.
Tell me what you decide.
Olivier
