### 2021-05-28 #55 @khinsen
@pantale Thanks for willingness to review! I'd say go ahead, we will see how this works out. We will have to figure out one day how to deal with HPC submissions, so "now" looks like a good moment :-)
