### 2021-06-01 #55 @pantale
@khinsen Hello,
Difficult for me to evaluate properly the proposed paper, but are some remarks:

The authors take care in the introduction to clearly distinguish the notion of replication from the notion of reproduction, which is a plus for the positioning of the problematic of the article. The source code and the data of the original article are not available, so it is indeed a reproduction.
The authors took care to make available on github all the data and sources used for the production of the article.

Points of detail:
The authors use bibliographic references and footnotes in the text, with exactly the same formalism of recall for both types of reference, making it difficult to read. For example, on page 2, Li and Dong5 refer to the reference, while github1 refers to the footnote. These two types of notation should be distinguished.

Methodology:
On page 4, the authors state that in the original paper by Li and Dong, the method for calculating hydrodynamic power is not specified. They bring their own solution and specify that "the surface pressure is defined as the fluid pressure interpolated at a distance of 3%", what is then the influence of this value on the reproducibility of the result, and why such a choice?

The hardware configuration used is not clear enough for me and differs between the abstract and section 3. The number of CPUs and Nvidia GPUs involved in the simulations is not clear from the article, as well as the RAM and GPU memory size.

Overall assessment:
Even though I am not a fluid mechanics specialist myself (my field of application is solid mechanics), I appreciated the detailed content of the article and the efforts provided by the authors to give as much detail as possible of the implementation and the different tracks allowing again to reproduce these simulations. The notable efforts concerning the provision of Docker images, simulation data, python scripts allowing the generation of figures in the article is a notable plus.

So without being a specialist in this type of numerical simulation, I think we can accept the proposed article with minor revisions for publication in ReScience C.

Olivier
