### 2021-06-04 #55 @khinsen
@mesnardo Something I can't find in your paper is the minimal hardware requirements to (1) run your Docker/singularity image or (2) run the whole code from source code, i.e. including a rebuild of the image. You do say "We used computational nodes with Dual 20-Core 3.70GHz Intel Xeon Gold 6148 processors and NVIDIA V100 GPU devices." but that doesn't mean that an exactly identical machine is required for replication. Would any recent Intel-based machine be OK? Are the GPUs required? Is the specifc type of GPU required?
