### 2021-06-15 #55 @mesnardo
@khinsen Thank you for the feedback.
We will add more details in the manuscript (or in the README of the application repository).

Although we have not tried it, I believe the Singularity image should run on a more recent Intel-based machine.
If it does not, we at least made the Dockerfiles and Singularity recipes available, so that one could adapt/re-build the images:

* [Dockerfile](https://github.com/barbagroup/petibm-rollingpitching/blob/master/docker/Dockerfile-petibm0.5.1-xenial) used to build the Docker image of the PetIBM-based application
  * It uses `barbagroup/petibm:0.5.1-GPU-OpenMPI-xenial-devel` as it based image (created from this [Dockerfile](https://github.com/barbagroup/petibm-recipes/blob/master/docker/Dockerfile-0.5.1-GPU-OpenMPI-xenial-devel))
  * The base image compiles and installs PetIBM and its dependencies, including NVIDIA AmgX (for solving systems on GPUs)
* [Singularity recipe](https://github.com/barbagroup/petibm-rollingpitching/blob/master/singularity/Singularity.petibm0.5.1-xenial) to convert Docker image into a Singularity one

The Docker image contains a version of PetIBM that enables the use of GPUs.
However, if no GPU devices are available on the host, one should still be able to use the same image and solve all iterative solvers on CPUs (with PETSc).
(We only used GPUs to solve the pressure Poisson system in our simulations; the two other systems were solved on CPUs.)

The Docker image contains a version of AmgX that was compiled to target GPUs with a compute capability of 3.5 (Kepler - Tesla K20, K40), 3.7 (Kepler - Tesla K80), 6.0 (Pascal - Tesla P100), and 7.0 (Volta - Tesla V100).

@rougier We would like to wait for the second review before making changes to the manuscript and replying to the reviewers.
