### 2021-06-15 #55 @chengyu1018
@khinsen

I have read the manuscript entitled, “[Re] Three-dimensional wake topology and propulsive performance of low-aspect-ratio pitching-rolling plates.” The authors successfully replicated the simulation results of the original study (Li & Dong, Physics of Fluids, 2016) with a different CFD solver. Although the force coefficient peaks showed a slightly lower magnitude than the original paper, the most important characteristic features were well captured, such as “double-C” -shaped vortex structures and force double peaks. It is worth noting that the “double-C”-shaped vortex structures were first reported by Li & Dong in the literature. Best of the reviewer’s knowledge, there have been no experimental studies that successfully observed this structure due to the complex wake structures generated by the bio-inspired pitching-rolling motion. So, the reviewer believes that the current Re-Science CFD study provides important evidence on the existence of such a unique wake structure since the numerical treatment between the original article and the Re-Science paper are very different. In addition, the overall observation on the effects of Reynolds number, Strouhal number, Aspect-ratio are consistent with the original article. I am not surprised by the difference in the force coefficient between the two studies. Depending on the numerical schemes used in different CFD solvers, some commercial/open-source software packages (e.g., Ansys Fluent) may introduce artificial viscosity for better convergence reasons.

The overall contents are well written and accessible to the broad readership of ReScience C. I only have one minor suggestion. It would be nice if the authors could consider adding a plot to demonstrate the mesh used in the current Re-Science study (like Fig.3 in Li & Dong’s paper).
