### 2021-06-19 #55 @khinsen
@mesnardo Since the reviewers didn't comment on the reproducibility, I am looking at this myself. At least theoretically (inspecting the files), and running as much as I can with the machines I have at my disposal.

I started with the easiest part: re-running your "repro packs". Using Docker for macOS, I can follow your instructions and I get freshly generated figures. However, only the line plots look the same. The 3D visualizations are not rendered at all, the figures contain nothing but the arrows and lines, often just the coordinate axes in the corner. One possible explanation is that the Docker image is made for NVIDIA graphics cards, which I don't have in my computer. If that's indeed the cause, I do find it worrying from a reproducibility perspective. If the code requires hardware (or systems software) that is not present, it should fail with an error message, instead of silently producing wrong output.

Next, I will try rebuilding the Docker and Singularity images. The instructions for that seem clear. The remaining step would be re-running the simulations themselves, which I might not be able to do. The instructions look clear enough, but I am not sure to be able to find suitable hardware.
