### 2021-06-21 #55 @mesnardo
@khinsen Thank you for attempting to build the image and opening an issue on the repo.
I'm away from keyboard this week, but will get back to work next Monday and investigate the issues about the 3D visualization and the Docker image.
We will also address the feedback from the reviewers.
