### 2021-06-23 #55 @khinsen
@mesnardo Here's another question for you (for when you are back, it's not urgent!), more about improving ReScience C than about your submission.

[Over here](https://github.com/ReScience/ReScience/issues/104#issuecomment-866317930), @benureau proposes a way to improve the reviewing process for HPC-style submissions. I'd like to have your view as an author on this idea. Would such a process be a reasonable burden on authors of future submissions? Note that I am *not* asking you to adopt this for the current submission, it's only about future submissions.
