### 2021-06-23 #55 @labarba
Note that the code that was used in our submission here was separately submitted, reviewed and published in JOSS. This means that reviewers installed the code, ran tests, and gave input to improve it. We have also made deposits with the primary data, as suggested by @benureau. And provide the secondary data (often smaller files, due to being summarized in some form) so the reproduction of figures is facilitated.

In a high-performance scenario, we need to shift our expectations of reproducibility being _exercised_ and view it as _prospective_ reproducibility, demonstrated by the documented process (how the research was conducted), and artifacts.

Related blog post from 2016: ["Why should I believe your supercomputing research?"](https://medium.com/@lorenaabarba/why-should-i-believe-your-supercomputing-research-a7cbf4cbc6b4)
