### 2021-06-23 #55 @khinsen
The JOSS paper is a good point, thanks @labarba! It would be helpful to point this out in the paper. The mere reference to a JOSS paper probably doesn't mean much for most readers (and reviewers). But it does indeed imply that someone has done an open review of the software, including installation and running the tests. We could even suggest to future authors to provide a link to a JOSS review if one exists for software they used. Let's start citing reviews!

And I fully agree of course that we need to shift our expectations - that's exactly what the discussion at https://github.com/ReScience/ReScience/issues/104 is about.
