### 2021-07-11 c03805b
Update readme with doi of container images
