### 2021-07-14 41eaabc
Add script to plot meshgrid and add generated figure
