### 2021-07-17 60eb292
Add pointer about hardware configuration and runtimes
