+++ runs/README.md
@@ -10,43 +10,46 @@ The present directory contains all the input files and post-processing scripts, 
 * `Re*_St*_AR*_psi*`: simulation folders of the parametric study; for example:
   * `Re200_St0.6_AR1.27_psi90`: simulation folder to compute the three-dimensional flow around a pitching and rolling flat plate with aspect ratio `AR=1.27` at Reynolds number `Re=200`, Strouhal number `St=0.6`, with a phase difference of `psi=90` degrees between the rolling and pitching motions.
 
-## Runtimes
+## Hardware configuration and run times
 
-Each simulation job was submitted to SLURM scheduling system on Pegasus (HPC cluster at GWU), using nodes of the `small-gpu` partition.
+Each simulation job was submitted to the SLURM scheduling system on Pegasus (HPC cluster at the George Washington University), using nodes of the `small-gpu` partition.
 
-Hardware configuration of small GPU nodes:
+Hardware configuration of the nodes on the `small-gpu` partition:
 
 * Dell PowerEdge R740 server
-* (2) NVIDIA Tesla V100 GPU
+* (2) NVIDIA Tesla V100 GPU (Memory size: 16GB)
 * Dual 20-Core 3.70GHz Intel Xeon Gold 6148 processors
 * 192GB of 2666MHz DDR4 ECC Register DRAM
 * 800 GB SSD onboard storage (used for boot and local scratch space)
 * Mellanox EDR Infiniband controller to 100GB fabric
 
+For each simulation, we used 20 CPUs and 2 GPU devices per node.
+Information about the number of nodes, the number of time steps computed, the mesh size, and the run time for each simulation are displayed in the following table.
+
 | Simulation | # time steps | # grid cells (10^6) | # nodes | runtime (hr) |
 |:-|:-:|:-:|:-:|:-:|
-| `independence/run1` | 10000 | 5.1 | 1 | 3.0 |
-| `independence/run2` | 10000 | 11.5 | 1 | 7.4 |
-| `independence/run3` | 10000 | 21.1 | 2 | 8.4 |
-| `independence/run4` | 10000 | 72.3 | 6 | 23.7 |
-| `independence/run5` | 10000 | 21.1 | 2 | 10.8 |
-| `independence/run6` | 5000 | 21.1 | 2 | 5.1 |
-| `independence/run7` | 10000 | 21.1 | 2 | 12.8 |
-| `Re100_St0.6_AR1.27_psi90` | 10000 | 21.1 | 2 | 11.0 |
-| `Re200_St0.4_AR1.27_psi90` | 10000 | 21.1 | 2 | 9.8 |
-| `Re200_St0.6_AR1.27_psi100` | 10000 | 21.1 | 2 | 9.1 |
-| `Re200_St0.6_AR1.27_psi110` | 10000 | 21.1 | 2 | 8.1 |
-| `Re200_St0.6_AR1.27_psi120` | 10000 | 21.1 | 2 | 9.0 |
-| `Re200_St0.6_AR1.27_psi60` | 10000 | 21.1 | 2 | 8.1 |
[…]
