### 2021-07-18 7d0c03d
Add version of Docker and Singularity to README
