### 2021-07-21 2727b85
Add warning to README about mem requirement for post-processing
and add some instructions to launch and AWS EC2 instance and run the Docker container on it to generate the secondary data and figures.
