### 2021-07-21 142fb9e
Add note about slurm submission scripts
