### 2021-08-03 #55 @labarba
Dear editor and reviewers: we have made changes responding to reviewer comments in this merged PR
https://github.com/barbagroup/rescience-rollingpitching/pull/3
