### 2021-08-04 061f755
Add instruction to run Singularity container on AWS
