### 2021-08-04 a06bec9
Add section on host requirements
