### 2021-08-04 #55 @mesnardo
@khinsen, @pantale, @chengyu1018 Thank you for the valuable feedback you provided!
We addressed your comments/questions with modifications to the manuscript and to the code application repository:

* Changes made to the **application code** repo were merged to the main branch via PR barbagroup/petibm-rollingpitching#4
* Changes made to the **manuscript** repo were merged to the main branch via PR barbagroup/rescience-rollingpitching#3
* The updated PDF of the manuscript can be found on the branch `revision-2021.08.03` ([direct link](https://github.com/barbagroup/rescience-rollingpitching/blob/revision-2021.08.03/article.pdf))

(The revised manuscript also adds a thank you to the reviewers and editor in the "Acknowledgments" section; in barbagroup/rescience-rollingpitching@a5e4ef4.)
