### 2021-08-04 #55 @mesnardo
Replying to @chengyu1018

> I only have one minor suggestion. It would be nice if the authors could consider adding a plot to demonstrate the mesh used in the current Re-Science study (like Fig.3 in Li & Dong’s paper).

That's a great suggestion; it will give a visual sense to the readers on the type of grids used in this study.
We added a figure containing a sub-view of the mesh grid (for the base case) to the manuscript (Figure 1).
Figure added to manuscript in barbagroup/rescience-rollingpitching@b0501c6.
Script to generate the figure added to the application repo in barbagroup/petibm-rollingpitching@41eaabc.
