### 2021-08-04 #55 @mesnardo
Replying to @pantale

> The authors use bibliographic references and footnotes in the text, with exactly the same formalism of recall for both types of reference, making it difficult to read. For example, on page 2, Li and Dong5 refer to the reference, while github1 refers to the footnote. These two types of notation should be distinguished.

100% agree.
To distinguish between references and footnotes, we now use symbols as footnote markers.
The switch to symbols was done in barbagroup/rescience-rollingpitching@aad7e5d.

> On page 4, the authors state that in the original paper by Li and Dong, the method for calculating hydrodynamic power is not specified. They bring their own solution and specify that "the surface pressure is defined as the fluid pressure interpolated at a distance of 3%", what is then the influence of this value on the reproducibility of the result, and why such a choice?

We added to the manuscript a more detailed paragraph on the calculation of the hydrodynamic power.
In this paragraph, we explain why we choose to interpolate at a distance of 3% of the chord length and provide numerical comparison to when interpolating closer or further away from the boundary.
The revised paragraph was added to the manuscript in barbagroup/rescience-rollingpitching@db55870.

> The hardware configuration used is not clear enough for me and differs between the abstract and section 3. The number of CPUs and Nvidia GPUs involved in the simulations is not clear from the article, as well as the RAM and GPU memory size.

I may have not understood your comment about the difference in reporting the hardware configuration (between the abstract and section 3).
However, we agree we should provide more information about the hardware configuration used for the simulations.
In barbagroup/rescience-rollingpitching@d52b864, we added (to the manuscript) a pointer for the readers on where to find the hardware configuration for each simulation (in the README of the application repo).
The README of the application repo contains a table with details about the hardware used for each simulation.
The table was added in barbagroup/petibm-rollingpitching@60eb292.
