### 2021-08-04 #55 @mesnardo
Replying to @khinsen

> Something I can't find in your paper is the minimal hardware requirements to (1) run your Docker/singularity image or (2) run the whole code from source code, i.e. including a rebuild of the image. You do say "We used computational nodes with Dual 20-Core 3.70GHz Intel Xeon Gold 6148 processors and NVIDIA V100 GPU devices." but that doesn't mean that an exactly identical machine is required for replication. Would any recent Intel-based machine be OK? Are the GPUs required? Is the specific type of GPU required?

We added a section to the README of the code repository with the requirements (software/hardware) to run the Singularity container (barbagroup/petibm-rollingpitching@a06bec9).
And considering that most of us do not have free access to machines with enough memory, we also added some instructions to launch and configure an AWS EC2 instance to run the simulations (barbagroup/petibm-rollingpitching@061f755).

> The 3D visualizations are not rendered at all, the figures contain nothing but the arrows and lines, often just the coordinate axes in the corner. One possible explanation is that the Docker image is made for NVIDIA graphics cards, which I don't have in my computer. If that's indeed the cause, I do find it worrying from a reproducibility perspective. If the code requires hardware (or systems software) that is not present, it should fail with an error message, instead of silently producing wrong output.

Starting from the repro-packs downloaded from Zenodo, I tried to re-generate the figures on my laptop (8GB of memory).
The process also failed to render the 3D visualizations on my laptop.
Digging into this issue, it appears that 8GB of memory is not enough to compute the secondary data (such as the 3D vorticity field).
(Looking into the log, I saw that the process for the application `petibm-vorticity` got killed.)
I then used an AWS EC2 instance (`t3.2xlarge`, 32 GB of memory) to confirm I could correctly generates the figures.
(You do not need access to NVIDIA GPUs to generate the secondary data and figures.)
It appears that a machine with more than 16GB of memory is required to generate the figures.
In barbagroup/petibm-rollingpitching@2727b85, we added a warning (to the README) about the memory requirement for generating secondary data and generating the figures.
In this commit, we also added some instructions to launch an AWS EC2 instance and run the Docker container to generate the data and figures.
(Anyone with a AWS account should now be able to download the repro-packs from Zenodo, to download the Docker images from Zenodo, and to execute the Docker container.)

> The JOSS paper is a good point, thanks @labarba! It would be helpful to point this out in the paper. The mere reference to a JOSS paper probably doesn't mean much for most readers (and reviewers).

100% agree.
In barbagroup/rescience-rollingpitching@e7f754a, we highlighted the JOSS paper and what it means for a software to be accepted and published in JOSS.

> Update: I can't build the petibm container image following the provided instructions: barbagroup/petibm-recipes#3

Thanks again for opening an issue on barbagroup/petibm-recipes#3.
There, you raised very good points about lacking reproducible builds for the Docker images (`apt update`, relying on base images that could be altered, etc.).
Thus, we deposited to Zenodo the Docker and Singularity images used for this study [![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.5090342.svg)](https://doi.org/10.5281/zenodo.5090342).
In barbagroup/petibm-rollingpitching@c03805b, we added the DOI to the README of the application repo (as well as the version used for Docker and Singularity).
We are also now citing the Zenodo deposit in the manuscript (barbagroup/rescience-rollingpitching@533466a).
In addition to that, we added sentences to the last section to explain why re-building a Docker image may fail or may not be reproducible (barbagroup/rescience-rollingpitching@533466a).
