### 2021-08-04 #55 @khinsen
Thanks @mesnardo for those very useful and relevant additions to your code and paper, and also for summarizing them here with precise references so that it's easy for us to find them. With the right tools and the right attitude, reviewing papers can actually be a nice experience!

@chengyu1018 and @pantale: please have a look at these changes and say if you are happy with them.
