### 2021-08-04 #55 @khinsen
@mesnardo As for the changes prompted by my own comments, they look very good. Providing AWS instructions is an excellent idea for helping others build on your work, as is archiving the container images.  I am not much of a fan of that approach in general, but as a workaround for difficult situations, that's fine with me.

The memory requirements do indeed explain my failure to reproduce the repro-pack. My laptop has 8GB, meaning that the virtual machine that Docker images are run with (under macOS) gets even less.
