### 2021-08-19 #55 @labarba
Sincere thank you to both reviewers and the editor for their contribution to this paper's publication in ReScience.
