### 2021-08-20 #55 @khinsen
Thanks @chengyu1018 and @pantale for reviewing this paper, which is now officially accepted!
And thanks to @mesnardo and @labarba for providing ReScience with this interesting challenge to its reviewing procedure.

I'll start with the technicalities of publication, and it is possible that I will request help from @mesnardo. Stay tuned.
