### 2021-08-23 #55 @khinsen
@mesnardo I tried to recompile your LaTeX file after updating the metadata. It works but I get suspicious warnings about multiply defined labels:
```
LaTeX Warning: Label `fig:independence_force_coefficients_dx' multiply defined.
LaTeX Warning: Label `fig:independence_force_coefficients_dx@cref' multiply def
ined.
```
Could you please fix this, or confirm that I can safely ignore the warnings?
