### 2021-08-23 #55 @khinsen
@mesnardo Alternatively, accept https://github.com/barbagroup/rescience-rollingpitching/pull/4 and generate the final PDF yourself!
