### 2021-08-23 #55 @mesnardo
@khinsen Thank you for the PR. I merged it and generated the PDF (after removing the duplicated figure, which was causing the warning). The PDF can be found [here](https://github.com/barbagroup/rescience-rollingpitching/blob/accepted-2021.08.23/article.pdf).
