### 2021-08-23 #55 @khinsen
The article has been published: http://doi.org/10.5281/zenodo.5234931. It will be listed on the ReScience Web site after the next update.
