### 2021-08-25 #55 @labarba
I'm a little confused by the footer indicating volume and issue. Is it Volume 7 issue 1 ? What do you use in ReScience for "page number"? Online-only journal generally use some article id for that.
